refactor: View 클래스로 차수별 실행 결과 출력 메소드 분리
