+++ src/App.js
@@ -8,7 +8,6 @@ class App {
     const {carNames, tryCount} = await this.readLineGameInputs();
 
     const game = new RacingGame();
-
     game.getInitialBoard(carNames);
 
     const getRacingRoundResult = (array, count) => {
@@ -29,18 +28,7 @@ class App {
     const scoreBoard = game.getScoreBoard();
     const racingRoundResult = getRacingRoundResult(scoreBoard, tryCount);
 
-    View.printNewLine();
-    MissionUtils.Console.print(`실행 결과`);
-
-    Utils.range(tryCount).forEach((_, index) => {
-      racingRoundResult.forEach((count) => {
-        const recordsRound = count.records.slice(0, index + 1).filter(Boolean).length;
-
-        const scoreChangeToDash = Utils.range(recordsRound).map(() => '-').join('');
-        MissionUtils.Console.print(`${count.name} : ${scoreChangeToDash}`);
-      });
-      View.printNewLine();
-    });
+    View.printRoundScore(tryCount, racingRoundResult);
 
     const finalWinner = game.getWinner(racingRoundResult);
     View.printFinalWinner(finalWinner);
